added some tests for the view models

+++ BS.Web/Scripts/App/Models/Profile.js
@@ -21,24 +21,39 @@
 
         ko.mapping.fromJS(data, mapping, self);
 
+        self.getBookFromReadingList = function (id) {
+            var index = self.ReadingList.mappedIndexOf({ Id: id });
+            return self.ReadingList()[index];
+        }
+
         self.isBookOnReadingList = function (id) {
             return self.ReadingList.mappedIndexOf({ Id: id }) != -1;
         };
 
-        self.addBookToReadingList = function (id) {
-            $.ajax({
-                type: "POST",
-                url: "api/" + self.Id + "/books",
-                data: { Id: id },
-                dataType: "json",
-                success: function (data) {
-                    self.ReadingList.push(data);
-                },
-                error: function (req, status, error) { debugger; }
-            });
+        self.addBookToReadingList = function (id, rating) {
+            var req = {};
+            if (!self.isBookOnReadingList(id)) {
+                req = $.ajax({
+                    type: "POST",
+                    url: "api/" + self.Id() + "/" + id(),
+                    data: { Rating: rating },
+                    dataType: "json"
+                }).success(function (res) { self.ReadingList.push(new BS.Book(res, self)); });
+            } else {
+                req = $.ajax({
+                    type: "PUT",
+                    url: "api/" + self.Id() + "/" + id(),
+                    data: { Rating: rating },
+                    dataType: "json"
+                }).success(function (res) { self.getBookFromReadingList(id).Rating(rating); });
+            }
+
+            req.error(function (res) { debugger; });
         }
 
         self.removeBookFromReadingList = function (id) {
+            if (!self.isBookOnReadingList(id)) return;
+
             $.ajax({
                 type: "DELETE",
                 url: "api/" + self.Id() + "/" + id(),
